Extract connection setup
