+++ src/common/framework/mysql/database.ts
@@ -1,4 +1,5 @@
 import * as mysql from 'mysql2';
+import { config } from '../config/config';
 
 export const query = async (
   connection: mysql.Connection | mysql.Pool,
@@ -12,4 +13,22 @@ export const query = async (
       resolve(rows);
     });
   });
+};
+
+export const getConnection = (): mysql.Connection => {
+  const configuration = config();
+  const connection: mysql.Connection = mysql.createConnection({
+    host: configuration.mesDatabaseHostname,
+    database: configuration.mesDatabaseName,
+    user: configuration.mesDatabaseUsername,
+    password: configuration.mesDatabasePassword,
+    charset: 'UTF8_GENERAL_CI',
+    ssl: process.env.TESTING_MODE ? null : 'Amazon RDS',
+    authSwitchHandler(data: any, cb: any) {
+      if (data.pluginName === 'mysql_clear_password') {
+        cb(null, Buffer.from(`${configuration.mesDatabasePassword}\0`));
+      }
+    },
+  });
+  return connection;
 };
