+++ src/functions/postResult/framework/handler.ts
@@ -6,7 +6,7 @@ import { decompressTestResult } from '../application/decompression-service';
 import { StandardCarTestCATBSchema } from '@dvsa/mes-test-schema/categories/B';
 import { saveTestResult } from '../application/save-result-service';
 import { TestResultDecompressionError } from '../domain/errors/test-result-decompression-error';
-import { bootstrapConfig } from './config/config';
+import { bootstrapConfig } from '../../../common/framework/config/config';
 
 export async function handler(event: APIGatewayProxyEvent, fnCtx: Context): Promise<Response> {
   await bootstrapConfig();
